Update Rotating Num.java
You are given two numbers n and k.  You are required to rotate n, k times to the right. 
If k is positive, rotate to the right i.e. remove right-most digit and make it leftmost. 
Do the reverse for negative value of k. 

Also k can have an absolute value larger than number of digits in n.


Note - Assume that the number of rotations will not cause leading 0's in the result. e.g. such an input will not be given
   n = 12340056
   k = 3
   r = 05612340

Sample Input
562984
2

Sample Output
845629